refactor: users services follow files architecture
