+++ src/services/users/confirmEmail/__docs__/get.yaml
